+++ src/engine/ui.c
@@ -10,11 +10,11 @@ KeyInputs keyInputs;
  * Create the window for drawing according to the available program
  * configuration.
  */
-static void UI_CreateWindow(const int width, const int height)
+static void UI_CreateWindow(const uint32_t width, const uint32_t height)
 {
-	const int windowInitFlags = SDL_WINDOW_OPENGL;
+	const uint32_t windowInitFlags = SDL_WINDOW_OPENGL;
 	window = SDL_CreateWindow("Powerblocks (knell)", SDL_WINDOWPOS_CENTERED,
-			SDL_WINDOWPOS_CENTERED, width, height, windowInitFlags);
+			SDL_WINDOWPOS_CENTERED, (int)width, (int)height, windowInitFlags);
 	if (window == NULL) {
 		SDL_Quit();
 		KN_FATAL_ERROR("Unable to create SDL windows\n");
